New project MES/requirements_mongo.txt
Flask>=3.0
Flask-Login>=0.6
mongoengine>=0.29
python-dotenv>=1.0
Werkzeug>=3.0